fix(build): keep connector SSRF fetch out of the client bundle
The connectors SSRF fix routed s3/gitlab/sentry/obsidian through
secureFetchWithRetry, which transitively imports input-validation.server
(and its Node-only `dns/promises`). connectors/registry.ts is imported by
client components for connector metadata, so the connector sync code —
which only ever runs in server API routes — gets pulled into the client
bundle, and Turbopack fails to resolve `dns/promises` (no browser shim).

- Move secureFetchWithRetry into a dedicated `secure-fetch.server` module so
  the shared documents/utils stays client-safe; connectors import from there.
- Add a browser-only `turbopack.resolveAlias` stub for `dns`/`dns/promises`
  (the documented Next 16 remedy). Server bundles keep the real module, so
  SSRF validation is unaffected — only the never-executed client copy is stubbed.

Verified with a full `next build` (compiles successfully, no module errors).
