+++ apps/sim/next.config.ts
@@ -91,6 +91,17 @@ const nextConfig: NextConfig = {
       './lib/execution/sandbox/bundles/*.cjs',
     ],
   },
+  turbopack: {
+    resolveAlias: {
+      // `dns/promises` has no browser shim. Server-only connector fetch logic
+      // (which imports `input-validation.server`) is statically reachable from
+      // the client bundle via the connector registry, but never runs there.
+      // Stub it for the browser only; the server keeps the real module so SSRF
+      // validation is unaffected.
+      'dns/promises': { browser: './lib/core/security/empty-node-fallback.browser.ts' },
+      dns: { browser: './lib/core/security/empty-node-fallback.browser.ts' },
+    },
+  },
   experimental: {
     optimizeCss: true,
     preloadEntriesOnStart: false,
